Consolidate sPCA functions in single module
